Bryan: Separated Date/Time; Moved time to header
Now time is kept in top left corner, whereas date can be removed if
need be. By adding fluid modifier, footer now takes full space
