+++ js/src/personal/bryan/Bryan.page.tsx
@@ -4,6 +4,7 @@ import { theme } from '@/personal/bryan/theme.ts'
 import { NavBar } from '@/personal/bryan/NavBar.tsx'
 import classes from './Bryan.module.css'
 import { Socials } from './Socials.tsx'
+import { TimeDisplay } from '@/personal/bryan/TimeDisplay.tsx'
 
 export function BryanPage() {
   //const [currentPage, setCurrentPage] = useState(Page.Home)
@@ -12,6 +13,7 @@ export function BryanPage() {
     <MantineProvider theme={theme}>
       <NavBar />
       <div className={classes.container}>
+        <TimeDisplay />
         <Outlet />
         <Socials />
       </div>
